Add back off throttling on message failures.

+++ gnsq/reader.py
@@ -2,6 +2,7 @@ import logging
 import random
 import gevent
 import blinker
+import time
 
 from itertools import cycle
 from collections import defaultdict
@@ -9,6 +10,7 @@ from collections import defaultdict
 from .lookupd import Lookupd
 from .nsqd import Nsqd
 from .backofftimer import BackoffTimer
+from .states import INIT, RUNNING, BACKOFF, THROTTLED, CLOSED
 
 from .errors import (
     NSQException,
@@ -30,7 +32,7 @@ class Reader(object):
         requeue_delay=0,
         lookupd_poll_interval=60,
         lookupd_poll_jitter=0.3,
-        low_rdy_idle_timeout=10,
+        low_ready_idle_timeout=10,
         max_backoff_duration=128,
     ):
         if not nsqd_tcp_addresses and not lookupd_http_addresses:
@@ -63,16 +65,23 @@ class Reader(object):
         self.requeue_delay = requeue_delay
         self.lookupd_poll_interval = lookupd_poll_interval
         self.lookupd_poll_jitter = lookupd_poll_jitter
+        self.low_ready_idle_timeout = low_ready_idle_timeout
         self.max_backoff_duration = max_backoff_duration
+
+        self.total_ready_count = 0
+        self._need_ready_redistributed = False
+        self.last_random_ready = time.time()
 
         self.logger = logging.getLogger(__name__)
         self.backofftimers = defaultdict(self.create_backoff)
+        self.backoff = self.create_backoff()
 
         self.on_response = blinker.Signal()
         self.on_error = blinker.Signal()
         self.on_message = blinker.Signal()
         self.on_finish = blinker.Signal()
         self.on_requeue = blinker.Signal()
+        self.on_giving_up = blinker.Signal()
         self.on_exception = blinker.Signal()
 
         self.conns = set()
@@ -80,24 +89,33 @@ class Reader(object):
 
         self.workers = []
         self.conn_workers = {}
-        self.closed = True
+        self.state = INIT
+
+    @property
+    def is_running(self):
+        self.state in (RUNNING, BACKOFF, THROTTLED)
 
     def start(self, block=True):
-        self.closed = False
+        if not self.start == INIT:
+            return
+
+        self.state = RUNNING
         self.query_nsqd()
 
         if self.lookupds:
             self.query_lookupd()
             self.workers.append(gevent.spawn(self._poll_lookupd))
 
-        # TODO
-        # self.workers.append(gevent.spawn(self._redistribute_rdy_state))
+        self.workers.append(gevent.spawn(self._poll_ready))
 
         if block:
             self.join()
 
     def close(self, block=True):
-        self.closed = True
+        if not self.is_running:
+            return
+
+        self.state = CLOSED
 
         for worker in self.workers:
             worker.kill(block=block)
@@ -112,14 +130,24 @@ class Reader(object):
         gevent.joinall(self.workers, timeout, raise_error)
         gevent.joinall(self.conn_workers.values(), timeout, raise_error)
 
+    def is_starved(self):
+        for conn in self.conns:
+            if conn.in_flight >= max(conn.last_ready * 0.85, 1):
+                return True
+        return False
+
+    @property
     def connection_max_in_flight(self):
         return max(1, self.max_in_flight / max(1, len(self.conns)))
 
-    def is_starved(self):
-        for conn in self.conns:
-            if conn.in_flight > 0 and conn.in_flight >= (conn.last_rdy * 0.85):
-                return True
-        return False
+    @property
+    def total_ready_count(self):
+        return sum(c.ready_count for c in self.conns)
+
+    def send_ready(self, conn, count):
+        if (self.total_ready_count + count) > self.max_in_flight:
+            return
+        conn.ready(count)
 
     def query_nsqd(self):
         self.logger.debug('querying nsqd...')
@@ -157,8 +185,92 @@ class Reader(object):
 
         while True:
             gevent.sleep(self.lookupd_poll_interval)
-            self.query_nsqd()
             self.query_lookupd()
+
+    def _poll_ready(self):
+        while True:
+            gevent.sleep(5)
+            if not self.need_ready_redistributed:
+                continue
+            self.redistribute_ready_state()
+
+    @property
+    def need_ready_redistributed(self):
+        if self.state == BACKOFF:
+            return False
+
+        if self._need_ready_redistributed:
+            return True
+
+        if len(self.conns) > self.max_in_flight:
+            return True
+
+        if self.state == THROTTLED and len(self.conns) > 1:
+            return True
+
+    @need_ready_redistributed.setter
+    def need_ready_redistributed(self, value):
+        self._need_ready_redistributed = value
+
+    def redistribute_ready_state(self):
+        self.need_ready_redistributed = False
+
+        # first set RDY 0 to all connections that have not received a message
+        # within a configurable timeframe (low_ready_idle_timeout).
+        for conn in self.conns:
+            if conn.ready_count == 0:
+                continue
+
+            if (time.time() - conn.last_message) < self.low_ready_idle_timeout:
+                continue
+
+            msg = '[{}] idle connection, giving up RDY count'.format(conn)
+            self.logger.info(msg)
+
+            conn.ready(0)
+
+        if self.state == THROTTLED:
+            max_in_flight = 1 - self.total_ready
+        else:
+            max_in_flight = self.max_in_flight - self.total_ready
+
+        if max_in_flight <= 0:
+            return
+
+        # randomly walk the list of possible connections and send RDY 1 (up to
+        # our calculate "max_in_flight").  We only need to send RDY 1 because in
+        # both cases described above your per connection RDY count would never
+        # be higher.
+        #
+        # We also don't attempt to avoid the connections who previously might
+        # have had RDY 1 because it would be overly complicated and not actually
+        # worth it (ie. given enough redistribution rounds it doesn't matter).
+        conns = list(self.conns)
+        conns = random.sample(conns, min(max_in_flight, len(self.conn)))
+
+        for conn in conns:
+            self.logger.info('[{}] redistributing RDY'.format(conn))
+            self.send_ready(conn, 1)
+
+    def random_ready_conn(self, conn):
+        # if all connections aren't getting RDY
+        # occsionally randomize which connection gets RDY
+        if len(self.conns) <= self.max_in_flight:
+            return conn
+
+        if (time.time() - self.last_random_ready) < 30:
+            return conn
+
+        self.last_random_ready = time.time()
+        return random.choice(c for c in self.conns if not c.ready_count)
+
+    def update_ready(self, conn):
+        if self.state in (BACKOFF, THROTTLED):
+            return
+
+        conn = self.random_ready_conn(conn)
+        if conn.ready_count < max(conn.last_ready * 0.25, 2):
+            self.send_ready(conn, self.connection_max_in_flight)
 
     def random_connection(self):
         if not self.conns:
@@ -173,7 +285,7 @@ class Reader(object):
         conn.publish(topic, message)
 
     def connect_to_nsqd(self, conn):
-        if self.closed:
+        if not self.is_running:
             return
 
         if conn in self.conns:
@@ -198,17 +310,17 @@ class Reader(object):
             conn.connect()
             conn.identify()
 
-            if conn.max_rdy_count < self.max_in_flight:
+            if conn.max_ready_count < self.max_in_flight:
                 self.logger.warning(' '.join([
                     '[{}] max RDY count {} < reader max in flight {},',
                     'truncation possible'
-                ]).format(conn, conn.max_rdy_count, self.max_in_flight))
+                ]).format(conn, conn.max_ready_count, self.max_in_flight))
 
             conn.subscribe(self.topic, self.channel)
 
-            # Send RDY 1 if we're the first connection
-            if not self.conns:
-                conn.ready(1)
+            # Send RDY 1 if we're not backing off or we're the first connection
+            if self.state != BACKOFF or not self.conns:
+                self.send_ready(conn, 1)
 
         except NSQException as error:
             msg = '[{}] connection failed ({!r})'.format(conn, error)
@@ -220,7 +332,7 @@ class Reader(object):
             self.pending.remove(conn)
 
         # Check if we've closed since we started
-        if self.closed:
+        if not self.is_running:
             conn.close_stream()
             return
 
@@ -250,6 +362,9 @@ class Reader(object):
         self.conn_workers.pop(conn, None)
         conn.close_stream()
 
+        if conn.ready_count:
+            self.need_ready_redistributed = True
+
         if str(conn) not in self.nsqd_tcp_addresses:
             return
 
@@ -267,10 +382,12 @@ class Reader(object):
 
     def handle_message(self, conn, message):
         self.logger.debug('[{}] got message: {}'.format(conn, message.id))
+        self.update_ready(conn)
 
         if self.max_tries and message.attempts > self.max_tries:
             msg = "giving up on message '{}' after max tries {}"
             self.logger.warning(msg.format(message.id, self.max_tries))
+            self.on_giving_up.send(self, conn, message)
             return message.finish()
 
         try:
@@ -289,23 +406,58 @@ class Reader(object):
 
         message.requeue(self.requeue_delay)
 
-    def update_ready(self, conn):
-        max_in_flight = self.connection_max_in_flight()
-        if conn.ready_count < (0.25 * max_in_flight):
-            conn.ready(max_in_flight)
-
     def handle_finish(self, conn, message_id):
         self.logger.debug('[{}] finished message: {}'.format(conn, message_id))
+        self.backoff.success()
+        self.handle_backoff()
         self.on_finish.send(self, conn=conn, message_id=message_id)
-        self.update_ready(conn)
 
     def handle_requeue(self, conn, message_id, timeout):
         msg = '[{}] requeued message: {} ({})'
         self.logger.debug(msg.format(conn, message_id, timeout))
+        self.backoff.failure()
+        self.handle_backoff()
         self.on_requeue.send(
             self,
             conn=conn,
             message_id=message_id,
             timeout=timeout
         )
-        self.update_ready(conn)
+
+    def handle_backoff(self):
+        if self.state == BACKOFF:
+            return
+
+        if self.state == THROTTLED and self.backoff.is_reset():
+            return self.complete_backoff()
+
+        if self.backoff.is_reset():
+            return
+
+        self.start_backoff()
+
+    def start_backoff(self):
+        self.state = BACKOFF
+
+        for conn in self.conns:
+            conn.ready(0)
+
+        interval = self.backoff.get_interval()
+        self.logger.info('backing off for {} seconds'.format(interval))
+        gevent.sleep(interval)
+
+        self.state = THROTTLED
+        if not self.conns:
+            return
+
+        conn = self.random_connection()
+        self.logger.info('[{}] testing backoff state with RDY 1'.format(conn))
+        self.send_ready(conn, 1)
+
+    def complete_backoff(self):
+        self.state = RUNNING
+        self.logger.info('backoff complete, resuming normal operation')
+
+        count = self.connection_max_in_flight
+        for conn in self.conns.values():
+            self.send_ready(conn, count)
